requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.12.1
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
lxml==6.0.2
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.1
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
protobuf==6.33.5
pycparser==3.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-docx==1.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
shiboken6==6.10.2
sounddevice==0.5.5
soundfile==0.13.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer-slim==0.21.2
typing_extensions==4.15.0
urllib3==2.6.3
